refactor: all_history as module
